backend/requirements.txt
Django
djangorestframework
gunicorn
psycopg2
django-filter
pytest
pytest-django


